Move code to alert_utils.py
